Modified mkimage to build RCW binaries
mkimage was modified to build a binary reset configuration word (RCW) file for Freescale PowerArchitecture controllers. CONFIG_FSL_RCW_BIN has to be defined and CONFIG_SYS_FSL_RCW points to the RCW configuration file that is the source of the image.
With this modification a RCW blob could be generated from the uboot source tree. This could e.g. be programmed to the NOR flash.