TQMT2081: vid: check uC register if setting voltage was successful
The microcontroller on the TQMT2081 module receives the desired VID
voltage via an I2C write to its VID_VALUE register.
After successfully changing the core voltage to the desired VID voltage
the uC sets the VID_VALUE register to 0xCC.
This is read back by the u-boot and considered as indicator that the
core voltage has been set correctly.

Signed-off-by: Stefan Lange <s.lange@gateware.de>