ctags: Minor changes to fix ctags output
At present ctags emits lines with unmatched quotes which means that the
output file is invalid. This is with exuberant-ctags version 5.9~svn201103
but I also see it with plain ctags. I am not sure that it is a bug though.

Make a few minor changes in the source code to fix this problem.

Signed-off-by: Simon Glass <sjg@chromium.org>
(cherry picked from commit 07bc873c73315854f52506777d9afb70c5cfedda)
Signed-off-by: Max Merchel <Max.Merchel@tq-group.com>