tqmt104x: fix IFC LAW mapping size
According to the documentation, the IFC mapping has a size of 128MiB,
not 256MiB. This would create a conflict with the ranges for BMAN and QMAN
if the configuration were actually valid. However
CONFIG_SYS_FLASH_BASE_PHYS (0xfe8000000) is not aligned to 256MiB.
Apparently the bits of the address are effectively masked to fix the
alignment, making the mapping actually start at 0xfe0000000, avoiding
the address conflict.

Fix the mapping size, so it matches the documentation, and the effective
configuration matches the code.

Signed-off-by: Matthias Schiffer <matthias.schiffer@tq-group.com>