tqmt104x: fix IFC TLB mapping
The TLB mapping had the wrong size, and all mappings must be aligned to
their size. Attempting to create a mapping of 256MiB at 0xe8000000
actually resulted in a mapping at 0xe0000000 instead (according to the
output of reginfo).

Unfortunately, the CPU does not support pages of size 128MiB, so we need
to create two mappings of 64MiB instead, shifting the indices of the
following TLB entries.

As the IFC mapping is required during early boot when booting from NOR
flash, and different flags are required in this phase, the flags of the
mapping are later fixed up in board_early_init_r(). Setting the final
flags right away in tlb_table for the second half results in significantly
slower boot, so we avoid it and extend the fixup to cover both halves
instead.

Signed-off-by: Matthias Schiffer <matthias.schiffer@tq-group.com>